// src/enums.ts
export enum RecordType {
    META_EOF = 0x0000,
    META_REALIZEPALETTE = 0x0035,
    META_SETPALENTRIES = 0x0037,
    META_SETBKMODE = 0x0102,
    META_SETMAPMODE = 0x0103,
    META_SETROP2 = 0x0104,
    META_SETRELABS = 0x0105,
    META_SETPOLYFILLMODE = 0x0106,
    META_SETSTRETCHBLTMODE = 0x0107,
    META_SETTEXTCHAREXTRA = 0x0108,
    META_RESTOREDC = 0x0127,
    META_RESIZEPALETTE = 0x0139,
    META_DIBCREATEPATTERNBRUSH = 0x0142,
    META_SETLAYOUT = 0x0149,
    META_SETBKCOLOR = 0x0201,
    META_SETTEXTCOLOR = 0x0209,
    META_OFFSETVIEWPORTORG = 0x0211,
    META_LINETO = 0x0213,
    META_MOVETO = 0x0214,
    META_OFFSETCLIPRGN = 0x0220,
    META_FILLREGION = 0x0228,
    META_SETMAPPERFLAGS = 0x0231,
    META_SELECTPALETTE = 0x0234,
    META_POLYGON = 0x0324,
    META_POLYLINE = 0x0325,
    META_SETTEXTJUSTIFICATION = 0x020A,
    META_SETWINDOWORG = 0x020B,
    META_SETWINDOWEXT = 0x020C,
    META_SETVIEWPORTORG = 0x020D,
    META_SETVIEWPORTEXT = 0x020E,
    META_OFFSETWINDOWORG = 0x020F,
    META_SCALEWINDOWEXT = 0x0410,
    META_SCALEVIEWPORTEXT = 0x0412,
    META_EXCLUDECLIPRECT = 0x0415,
    META_INTERSECTCLIPRECT = 0x0416,
    META_ELLIPSE = 0x0418,
    META_FLOODFILL = 0x0419,
    META_FRAMEREGION = 0x0429,
    META_ANIMATEPALETTE = 0x0436,
    META_TEXTOUT = 0x0521,
    META_POLYPOLYGON = 0x0538,
    META_EXTFLOODFILL = 0x0548,
    META_RECTANGLE = 0x041B,
    META_SETPIXEL = 0x041F,
    META_ROUNDRECT = 0x061C,
    META_PATBLT = 0x061D,
    META_SAVEDC = 0x001E,
    META_PIE = 0x081A,
    META_STRETCHBLT = 0x0B23,
    META_ESCAPE = 0x0626,
    META_INVERTREGION = 0x012A,
    META_PAINTREGION = 0x012B,
    META_SELECTCLIPREGION = 0x012C,
    META_SELECTOBJECT = 0x012D,
    META_SETTEXTALIGN = 0x012E,
    META_ARC = 0x0817,
    META_CHORD = 0x0830,
    META_BITBLT = 0x0922,
    META_EXTTEXTOUT = 0x0a32,
    META_SETDIBTODEV = 0x0d33,
    META_DIBBITBLT = 0x0940,
    META_DIBSTRETCHBLT = 0x0b41,
    META_STRETCHDIB = 0x0f43,
    META_DELETEOBJECT = 0x01f0,
    META_CREATEPALETTE = 0x00f7,
    META_CREATEPATTERNBRUSH = 0x01F9,
    META_CREATEPENINDIRECT = 0x02FA,
    META_CREATEFONTINDIRECT = 0x02FB,
    META_CREATEBRUSHINDIRECT = 0x02FC,
    META_CREATEREGION = 0x06FF
};
// ...
export enum MetafileEscapes {
    NEWFRAME = 0x0001,
    ABORTDOC = 0x0002,
    NEXTBAND = 0x0003,
    SETCOLORTABLE = 0x0004,
    GETCOLORTABLE = 0x0005,
    FLUSHOUT = 0x0006,
    DRAFTMODE = 0x0007,
    QUERYESCSUPPORT = 0x0008,
    SETABORTPROC = 0x0009,
    STARTDOC = 0x000A,
    ENDDOC = 0x000B,
    GETPHYSPAGESIZE = 0x000C,
    GETPRINTINGOFFSET = 0x000D,
    GETSCALINGFACTOR = 0x000E,
    META_ESCAPE_ENHANCED_METAFILE = 0x000F,
    SETPENWIDTH = 0x0010,
    SETCOPYCOUNT = 0x0011,
    SETPAPERSOURCE = 0x0012,
    PASSTHROUGH = 0x0013,
    GETTECHNOLOGY = 0x0014,
    SETLINECAP = 0x0015,
    SETLINEJOIN = 0x0016,
    SETMITERLIMIT = 0x0017,
    BANDINFO = 0x0018,
    DRAWPATTERNRECT = 0x0019,
    GETVECTORPENSIZE = 0x001A,
    GETVECTORBRUSHSIZE = 0x001B,
    ENABLEDUPLEX = 0x001C,
    GETSETPAPERBINS = 0x001D,
    GETSETPRINTORIENT = 0x001E,
    ENUMPAPERBINS = 0x001F,
    SETDIBSCALING = 0x0020,
    EPSPRINTING = 0x0021,
    ENUMPAPERMETRICS = 0x0022,
    GETSETPAPERMETRICS = 0x0023,
    POSTSCRIPT_DATA = 0x0025,
    POSTSCRIPT_IGNORE = 0x0026,
    GETDEVICEUNITS = 0x002A,
    GETEXTENDEDTEXTMETRICS = 0x0100,
    GETPAIRKERNTABLE = 0x0102,
    EXTTEXTOUT = 0x0200,
    GETFACENAME = 0x0201,
    DOWNLOADFACE = 0x0202,
    METAFILE_DRIVER = 0x0801,
    QUERYDIBSUPPORT = 0x0C01,
    BEGIN_PATH = 0x1000,
    CLIP_TO_PATH = 0x1001,
    END_PATH = 0x1002,
    OPENCHANNEL = 0x100E,
    DOWNLOADHEADER = 0x100F,
    CLOSECHANNEL = 0x1010,
    POSTSCRIPT_PASSTHROUGH = 0x1013,
    ENCAPSULATED_POSTSCRIPT = 0x1014,
    POSTSCRIPT_IDENTIFY = 0x1015,
    POSTSCRIPT_INJECTION = 0x1016,
    CHECKJPEGFORMAT = 0x1017,
    CHECKPNGFORMAT = 0x1018,
    GET_PS_FEATURESETTING = 0x1019,
    MXDC_ESCAPE = 0x101A,
    SPCLPASSTHROUGH2 = 0x11D8
};